+++ archive/legacy/ui/__init__.py
@@ -1,3 +1,0 @@
-"""
-Pacote UI do SwitchPilot
-"""